move the credentials dialog into the manager implementation
